% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_asset.R
\name{check_data_asset}
\alias{check_data_asset}
\title{Check data asset}
\usage{
check_data_asset(data_assets, types, p_e, call = parent.frame())
}
\arguments{
\item{data_assets}{List: list of data assets to be checked}

\item{types}{List: list of types to be checked}

\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
None
}
\description{
Helper function to check data asset
Used in:
\itemize{
\item copy_from_dependency_list
\item grab_target_list
\item pipeline_import_for
\item mark_for_refresh
\item pipeline_run
}
}
\keyword{internal}
