% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{check_dependencies_not_listed}
\alias{check_dependencies_not_listed}
\title{Check dependencies used and not mentioned in object_relations}
\usage{
check_dependencies_not_listed(
  object_name,
  p_e = pipeline_env,
  test_mode = FALSE
)
}
\arguments{
\item{object_name}{Character. The name of the object.}

\item{p_e}{pipeline environment}
}
\value{
invisible NULL
}
\description{
From the initialized pipeline environment, \code{check_dependencies_not_listed} to check
whether any of the dependencies specified in the config files are used but not listed in \code{object_relations.yaml}
config. The location of the R script
is retrieved from the pipeline environment object \code{p_e}. The usage check is done by performing a string
search  within the R script file.
This is necessary because dependency relations may be missing, but the pipeline
can still succeed if the dependency is referenced in an object that was generated earlier.
}
\keyword{internal}
