% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_render_report.R
\name{check_extension_helper}
\alias{check_extension_helper}
\title{Check extension (helper)}
\usage{
check_extension_helper(filename, ext)
}
\arguments{
\item{filename}{string, the filename}

\item{ext, }{string, the extension to check for}
}
\value{
filename with extension
}
\description{
This function checks if the desired extension is present or not.
If the extension is not present, it concatenated it to the filenmae
}
\keyword{internal}
