% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_list}
\alias{check_list}
\title{Check list}
\usage{
check_list(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{list object: the list to be checked}

\item{...}{Additional arguments. Should be empty.}

\item{allow_null}{Boolean: option to set allowance for null.}

\item{arg}{Object: Argument name that gets fetched from argument}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
None
}
\description{
Helper function to check if input is in list.
used in \code{open_tbl_con}
}
\keyword{internal}
