% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_render_report.R
\name{check_path_helper}
\alias{check_path_helper}
\title{Check path (helper)}
\usage{
check_path_helper(path, file)
}
\arguments{
\item{path}{string, the path to the file}

\item{file}{string, the name of the file}
}
\value{
string, absolute, checked filepath of file
}
\description{
Check path (helper)
}
\keyword{internal}
