% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pipeline_init.R
\name{check_pipeline_init}
\alias{check_pipeline_init}
\title{Check pipeline init}
\usage{
check_pipeline_init(p_e = pipeline_env, call = parent.frame())
}
\arguments{
\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
call
}
\description{
Helper function to check pipeline used in:
\itemize{
\item grab_target_list
\item grab_current_drive_name
\item grab_target_mount
\item grab_run_timestamp
\item grab_object_table
\item pipeline_vis
\item pipeline_status
}
}
\keyword{internal}
