% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_target_list.R
\name{check_tags}
\alias{check_tags}
\title{Check tags}
\usage{
check_tags(tags, p_e, call = parent.frame())
}
\arguments{
\item{tags}{List: list of tags to check}

\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
None
}
\description{
Helper function used to check tags
}
\keyword{internal}
