% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{conf_check_empty}
\alias{conf_check_empty}
\title{check config empty}
\usage{
conf_check_empty(conf = conf, .call = parent.frame())
}
\arguments{
\item{conf}{loaded config file}

\item{.call}{Object: the parent environment from return_env, internal - do not use}
}
\description{
Helper function to check if config is empty
}
\keyword{internal}
