% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{conf_check_load_proj_paths}
\alias{conf_check_load_proj_paths}
\title{Check and load project paths}
\usage{
conf_check_load_proj_paths(f, .call = parent.frame())
}
\arguments{
\item{f}{Path: filepath}

\item{.call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
yaml
}
\description{
Helper function to read project paths
}
\keyword{internal}
