% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{conf_check_load_schedules}
\alias{conf_check_load_schedules}
\title{Check and load schedules}
\usage{
conf_check_load_schedules(f, .call = parent.frame())
}
\arguments{
\item{f}{Path: filepath}

\item{.call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
yaml
}
\description{
Helper function to read object schedules
}
\keyword{internal}
