% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{conf_create_dag}
\alias{conf_create_dag}
\title{Create Dag}
\usage{
conf_create_dag(conf_relations, all_objects, call = parent.frame())
}
\arguments{
\item{conf_relations}{yaml: loaded yaml containing object relations from configuration file}

\item{all_objects}{All objects, returned from \code{conf_get_all_objects}}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
gr: Graph object
}
\description{
Helper function to create DAGs
}
\keyword{internal}
