% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{conf_get_all_objects}
\alias{conf_get_all_objects}
\title{Get all objects}
\usage{
conf_get_all_objects(conf_relations, conf_definitions, .call = parent.frame())
}
\arguments{
\item{conf_relations}{yaml: loaded yaml containing object relations from configuration file}

\item{conf_definitions}{yaml: loaded yaml containing object definitions from configuration file}

\item{.call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
combined objects
}
\description{
Helper function to get all objects
}
\keyword{internal}
