% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{conf_register_object}
\alias{conf_register_object}
\title{Register object}
\usage{
conf_register_object(
  object_name,
  conf_paths,
  conf_definitions,
  conf_schedules,
  object_dag,
  is_custom_stamp,
  p_e = pipeline_env,
  .call = parent.frame()
)
}
\arguments{
\item{object_name}{String: name of the object}

\item{conf_paths}{yaml: loaded yaml containing object paths from configuration file}

\item{conf_definitions}{yaml: loaded yaml containing object definitions from configuration file}

\item{conf_schedules}{yaml: loaded yaml containing object schedules from configuration file}

\item{object_dag}{Dag: Directed Acyclic Graph object}

\item{p_e}{Object: package_environment, internal - do not use}

\item{.call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
conf_object
}
\description{
Helper function to register object
}
\keyword{internal}
