% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_target_list.R
\name{filter_scheduled_objects}
\alias{filter_scheduled_objects}
\title{Filter scheduled objects}
\usage{
filter_scheduled_objects(objects, p_e, call = parent.frame())
}
\arguments{
\item{objects}{Character: name of the object to get schedule from}

\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
Character with objects that have an active (or no) schedule
}
\description{
Helper function used to filter objects based on whether they have an active schedule
}
\keyword{internal}
