% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_dependency_list.R
\name{get_copy_to_dir_path}
\alias{get_copy_to_dir_path}
\title{Get the copy to dir path from object_definitions}
\usage{
get_copy_to_dir_path(
  data_asset_name,
  p_e = pipeline_env,
  call = parent.frame()
)
}
\arguments{
\item{data_asset_name}{Character.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\value{
string with filepath to copy to
}
\description{
Retrieve the given location under \code{copy_to_file_path} in object_definitions.yaml.
}
\seealso{
Other Copy functions:
\code{\link[=copy_asset]{copy_asset()}},
\code{\link[=get_copy_from_file_path]{get_copy_from_file_path()}}
}
\concept{Copy functions}
\keyword{internal}
