% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{get_depends_on_list}
\alias{get_depends_on_list}
\title{Get all dependencies for checks on config}
\usage{
get_depends_on_list(object_name)
}
\arguments{
\item{object_name}{Character.}
}
\value{
An array of dependencies for the config dependency check functions.
}
\description{
This function gets all dependencies as listed in object table
}
\keyword{internal}
