% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_target_list.R
\name{get_effective_schedule}
\alias{get_effective_schedule}
\title{Get effective schedule}
\usage{
get_effective_schedule(object_name, p_e, call = parent.frame())
}
\arguments{
\item{object_name}{Character: name of the object to get schedule from}

\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
schedule of an object
}
\description{
Helper function used to get schedule for an object
}
\keyword{internal}
