% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_sp_data_argcheck}
\alias{get_sp_data_argcheck}
\title{Argument check for get sp data}
\usage{
get_sp_data_argcheck(con_specs, call = parent.frame())
}
\arguments{
\item{con_specs}{List: list with connection specifications}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
checked arguments
}
\description{
Helper function that checks connection arguments
}
\keyword{internal}
