% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_sp_data_parse_date}
\alias{get_sp_data_parse_date}
\title{Parse date for get sp data}
\usage{
get_sp_data_parse_date(d, call = parent.frame())
}
\arguments{
\item{d}{Date: Date object that could be either character, Date or POSIXct format}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
formatted date
}
\description{
Helper function to correctly parse date objects
}
\keyword{internal}
