% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_current_drive_name.R
\name{grab_current_drive_name}
\alias{grab_current_drive_name}
\title{Grab the drive_name for the current EPI pipeline.}
\usage{
grab_current_drive_name(..., p_e = pipeline_env, call = parent.frame())
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\value{
A string in with the drive name
}
\description{
Part of the experimental pipeline functions.
}
\seealso{
Other Grab functions:
\code{\link[=grab_object_cache]{grab_object_cache()}},
\code{\link[=grab_object_table]{grab_object_table()}},
\code{\link[=grab_run_timestamp]{grab_run_timestamp()}},
\code{\link[=grab_target_list]{grab_target_list()}},
\code{\link[=grab_target_mount]{grab_target_mount()}}
}
\concept{Grab functions}
