% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_target_mount.R
\name{grab_target_mount}
\alias{grab_target_mount}
\title{Grab the mount for the current EPI pipeline.}
\usage{
grab_target_mount(..., p_e = pipeline_env, call = parent.frame())
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\value{
A string in the form of "rivm"
}
\description{
Grabs and returns the current mount for the EPI pipeline.
}
\seealso{
Other Grab functions:
\code{\link[=grab_current_drive_name]{grab_current_drive_name()}},
\code{\link[=grab_object_cache]{grab_object_cache()}},
\code{\link[=grab_object_table]{grab_object_table()}},
\code{\link[=grab_run_timestamp]{grab_run_timestamp()}},
\code{\link[=grab_target_list]{grab_target_list()}}
}
\concept{Grab functions}
