% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_import_for.R
\name{import_object}
\alias{import_object}
\title{Import object}
\usage{
import_object(data_asset_name, p_e, call = parent.frame())
}
\arguments{
\item{data_asset_name}{Character. Name of the data_asset (objects/other) that you want to import object for.}

\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\description{
Helper function that imports the object.
}
\keyword{internal}
