% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_init.R
\name{init_run_timestamp}
\alias{init_run_timestamp}
\title{Initialize run timestamp}
\usage{
init_run_timestamp(run_timestamp, p_e)
}
\arguments{
\item{run_timestamp}{string: String in timestamp format (YYYYMMDD_HHMM)}

\item{p_e}{Object: package_environment, internal - do not use}
}
\description{
Helper function that initializes a run timestamp
}
\keyword{internal}
