% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_init.R
\name{init_target_mount}
\alias{init_target_mount}
\title{Initialize target mount}
\usage{
init_target_mount(p_e)
}
\arguments{
\item{p_e}{Object: package_environment, internal - do not use}
}
\description{
Helper function used to initialize target mount
}
\keyword{internal}
