% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_target_list.R
\name{is_schedule_active}
\alias{is_schedule_active}
\title{Is schedule active}
\usage{
is_schedule_active(schedule, run_timestamp, call = parent.frame())
}
\arguments{
\item{schedule}{Character: name of the object to get schedule from}

\item{run_timestamp}{Character: A string in the form YYYMMDD_HHMM of the current run timestamp}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
boolean whether schedule is active or not
}
\description{
Helper function used to check whether a schedule is active
}
\keyword{internal}
