% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_for_refresh.R
\name{mark_for_refresh}
\alias{mark_for_refresh}
\title{Mark data_assets for refresh}
\usage{
mark_for_refresh(data_assets = NULL, ..., p_e = pipeline_env)
}
\arguments{
\item{data_assets}{Character. Which data_assets (objects/data/other) should be
refreshed? Defaults to all data_assets of the pipeline.}

\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}
}
\description{
Ensures that data_asset will be reimported by
\code{\link[=pipeline_import_for]{pipeline_import_for()}} or regenerated during \code{\link[=pipeline_run]{pipeline_run()}} (whichever
comes first).
}
\concept{Development functions}
