% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_con.R
\name{open_con}
\alias{open_con}
\title{open connection}
\usage{
open_con(con_specs)
}
\arguments{
\item{con_specs}{List. Connection arguments for the database. It should be a
named list with the names "name"/"con_args"/"tbl_args". con_args should be
a named character with entries "server"/"database"/"uid"/"pwd" and tbl_args
should be a named character with the entries
"dbcatalog"(optional)/dbschema"/"dbtable".}
}
\value{
con
}
\description{
Helper function to open connection used in:
}
\keyword{internal}
