% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{open_sp}
\alias{open_sp}
\title{open SP}
\usage{
open_sp(con, sp_name, start_date, end_date, infection_code)
}
\arguments{
\item{con}{Object: connection object}

\item{sp_name}{string: Name of the database}

\item{start_date}{Date: start date for the query}

\item{end_date}{Date: end date for the query}

\item{infection_code}{string: infection code}
}
\value{
data
}
\description{
open SP
}
\keyword{internal}
