% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tbl_con.R
\name{open_tbl}
\alias{open_tbl}
\title{open table}
\usage{
open_tbl(con_specs, con)
}
\arguments{
\item{con_specs}{List. Connection arguments for the database. It should be a
named list with the names "name"/"con_args"/"tbl_args". con_args should be
a named character with entries "server"/"database"/"uid"/"pwd" and tbl_args
should be a named character with the entries
"dbcatalog"(optional)/dbschema"/"dbtable".}

\item{con}{Connection, which is established with \code{open_con}}
}
\value{
tbl_db
}
\description{
Helper function to access data table (tbl)
}
\keyword{internal}
