% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_tbl_con.R
\name{open_tbl_con_argcheck}
\alias{open_tbl_con_argcheck}
\title{Open table connection argument check}
\usage{
open_tbl_con_argcheck(con_specs, .call = parent.frame())
}
\arguments{
\item{con_specs}{List. Connection arguments for the database. It should be a
named list with the names "name"/"con_args"/"tbl_args". con_args should be
a named character with entries "server"/"database"/"uid"/"pwd" and tbl_args
should be a named character with the entries
"dbcatalog"(optional)/dbschema"/"dbtable".}

\item{.call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
None
}
\description{
Helper function to check missing/wrong values for con or tbl
}
\keyword{internal}
