% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_import_for.R
\name{pipeline_import_for}
\alias{pipeline_import_for}
\title{pipeline import for}
\usage{
pipeline_import_for(
  data_assets,
  ...,
  p_e = pipeline_env,
  call = parent.frame()
)
}
\arguments{
\item{data_assets}{Character. Name of the data_assets (objects/other) that you want to import objects for.
Multiple asset names should passed in an array of the form 'c()'.}

\item{...}{Additional arguments. Should be empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\description{
Imports the dependencies (objects / data / other) given a
list of asset names. The imported objects will be loaded into the global
environment.
}
\seealso{
Other Pipeline functions:
\code{\link[=check_last_successful_run]{check_last_successful_run()}},
\code{\link[=check_most_recent_data]{check_most_recent_data()}},
\code{\link[=pipeline_init]{pipeline_init()}},
\code{\link[=pipeline_render_report]{pipeline_render_report()}},
\code{\link[=pipeline_run]{pipeline_run()}},
\code{\link[=pipeline_status]{pipeline_status()}},
\code{\link[=pipeline_vis]{pipeline_vis()}},
\code{\link[=setup_environment]{setup_environment()}}
}
\concept{Pipeline functions}
