% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_init.R
\name{pipeline_init}
\alias{pipeline_init}
\title{Initiate DARApipeline for an EPI pipeline}
\usage{
pipeline_init(
  run_timestamp = NULL,
  custom_stamp = NULL,
  ...,
  p_e = pipeline_env
)
}
\arguments{
\item{run_timestamp}{Character. run_timestamp in the format YYYMMDD_HHMM.
Defaults to the current day + time. Set this parameter to an old
run_timestamp to 'go back in time'.}

\item{custom_stamp}{Character. Advanced use only. Add custom text to timestamp.
Particularly useful for multi-threaded pipeline runs. This bypasses the date format checks.}

\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}
}
\description{
Initiates configuration of DARApipeline for use in an EPI pipeline. Does the following steps:
\enumerate{
\item Creates a run_timestamp
\item Initiates logging to log file
\item Interprets the configuration in config/
}

DARApipeline will keep track any parameters (such as the \code{run_timestamp} and the
configuration) in a hidden pipeline state environment.
}
\examples{
\dontrun{
pipeline_init()
}

}
\seealso{
\code{\link[=grab_object_table]{grab_object_table()}}.

Other Pipeline functions:
\code{\link[=check_last_successful_run]{check_last_successful_run()}},
\code{\link[=check_most_recent_data]{check_most_recent_data()}},
\code{\link[=pipeline_import_for]{pipeline_import_for()}},
\code{\link[=pipeline_render_report]{pipeline_render_report()}},
\code{\link[=pipeline_run]{pipeline_run()}},
\code{\link[=pipeline_status]{pipeline_status()}},
\code{\link[=pipeline_vis]{pipeline_vis()}},
\code{\link[=setup_environment]{setup_environment()}}
}
\concept{Pipeline functions}
