% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_render_report.R
\name{pipeline_render_report}
\alias{pipeline_render_report}
\title{DARA function that renders markdown files}
\usage{
pipeline_render_report(
  markdown_file,
  output_file,
  add_timestamp = TRUE,
  custom_markdown_path = "",
  custom_output_path = "",
  quiet = TRUE,
  ...
)
}
\arguments{
\item{markdown_file}{name of .Rmd file inside EPI 02_report folder to be rendered}

\item{output_file}{name of he .html output file. save location is dir_html}

\item{add_timestamp}{boolean to add timestamp to output file. default = TRUE}

\item{custom_markdown_path}{advanced use only. If not empty, this path overwrites the default markdown filepath}

\item{custom_output_path}{advanced use only. If not empty, this path overwrites the default save location for reports}

\item{quiet}{boolean used by rmarkdown::render to determine verbose usage.}

\item{...}{For internal use. Leave empty.}
}
\value{
None, renders markdown file
}
\description{
Runs rmarkdown::render() with extra functionality.
Default markdown filepath is inside the EPI /scripts/02_report folder
Default output filepath is \code{dir_html}, which is stated in the config file:
grab_object_table() \%>\% filter(tag == "html") \%>\% pull(dir_html)

Custom paths overide the default paths
}
\examples{
\dontrun{
# Note that the different examples could be used interchangeably!

# Render report with standard EPI pipeline locations
pipeline_render_report(markdown_file = "dummy_report.Rmd",
                       output_file = "dummy_report.html")

# Render report with extra parameters (e.g. port script variables to markdown)
pipeline_render_report(markdown_file = "dummy_report.Rmd,
                       output_file = "dummy_report.html",
                       params = list(cur_hosp = current_hospital))

# Render report with custom location without timestamp
pipeline_render_report(markdown_file = "dummy_report.Rmd",
                       ouput_file = "my_custom_report.html"
                       add_timestamp = FALSE,
                       custom_output_path = "output/extra_output")

# Render markdown file from custom location
pipeline_render_report(markdown_file = "my_custom_markdown_file.Rmd,
                      output = "dummy_report.html",
                      custom_markdown_path = "extra_reports/")
}
}
\seealso{
Other Pipeline functions:
\code{\link[=check_last_successful_run]{check_last_successful_run()}},
\code{\link[=check_most_recent_data]{check_most_recent_data()}},
\code{\link[=pipeline_import_for]{pipeline_import_for()}},
\code{\link[=pipeline_init]{pipeline_init()}},
\code{\link[=pipeline_run]{pipeline_run()}},
\code{\link[=pipeline_status]{pipeline_status()}},
\code{\link[=pipeline_vis]{pipeline_vis()}},
\code{\link[=setup_environment]{setup_environment()}}
}
\concept{Pipeline functions}
