% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_run.R
\name{pipeline_run}
\alias{pipeline_run}
\title{pipeline_run}
\usage{
pipeline_run(tags = NULL, objects = NULL, ..., p_e = pipeline_env)
}
\arguments{
\item{tags}{Character. Which tags should be in the list of targets? Defaults
to all tags if objects = NULL.}

\item{objects}{Character. Which objects should be in the list of targets?
Defaults to all objects if tags = NULL.}

\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}
}
\description{
Runs a EPI pipeline, given a set of tags/objects filters.

A pipeline consists of the following steps:
\enumerate{
\item Determine all the targets that needs to be generated.
\item Then for each target not already in memory:
\enumerate{
\item Source the script that generates it (which internally imports dependencies using \code{\link[=pipeline_import_for]{pipeline_import_for()}}).
\item Cache the object to the cache folder.
\item Save the object to the output folder.
}
}

All of the configuration of the pipeline should be defined inside the config folder
and loaded into R using \code{\link[=pipeline_init]{pipeline_init()}}.
}
\seealso{
\code{\link[=mark_for_refresh]{mark_for_refresh()}}
\code{\link[=grab_target_list]{grab_target_list()}}

Other Pipeline functions:
\code{\link[=check_last_successful_run]{check_last_successful_run()}},
\code{\link[=check_most_recent_data]{check_most_recent_data()}},
\code{\link[=pipeline_import_for]{pipeline_import_for()}},
\code{\link[=pipeline_init]{pipeline_init()}},
\code{\link[=pipeline_render_report]{pipeline_render_report()}},
\code{\link[=pipeline_status]{pipeline_status()}},
\code{\link[=pipeline_vis]{pipeline_vis()}},
\code{\link[=setup_environment]{setup_environment()}}
}
\concept{Pipeline functions}
