% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_status.R
\name{pipeline_status}
\alias{pipeline_status}
\title{Print a status overview of the EPI pipeline}
\usage{
pipeline_status(..., p_e = pipeline_env)
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}
}
\description{
This function prints the current status of the pipeline:
\enumerate{
\item Whether the pipeline has been initiated.
\item The initiated run_timestamp.
\item How many objects have been created so far.
\item The location of the log file.
}
}
\seealso{
Other Pipeline functions:
\code{\link[=check_last_successful_run]{check_last_successful_run()}},
\code{\link[=check_most_recent_data]{check_most_recent_data()}},
\code{\link[=pipeline_import_for]{pipeline_import_for()}},
\code{\link[=pipeline_init]{pipeline_init()}},
\code{\link[=pipeline_render_report]{pipeline_render_report()}},
\code{\link[=pipeline_run]{pipeline_run()}},
\code{\link[=pipeline_vis]{pipeline_vis()}},
\code{\link[=setup_environment]{setup_environment()}}
}
\concept{Pipeline functions}
