% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_vis.R
\name{pipeline_vis}
\alias{pipeline_vis}
\title{Visualizes object creation as a graph.}
\usage{
pipeline_vis(..., p_e = pipeline_env)
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}
}
\description{
Visualizes the network defined by the object_relations yaml config file using
the \href{https://datastorm-open.github.io/visNetwork/}{visNetwork} package. This
function is meant for interactive use to view the structure of an EPI report
pipeline and can be very useful for inspecting the dependencies of your
objects. The pipeline_vis() function can maximally visualize 18 tags.
}
\seealso{
Other Pipeline functions:
\code{\link[=check_last_successful_run]{check_last_successful_run()}},
\code{\link[=check_most_recent_data]{check_most_recent_data()}},
\code{\link[=pipeline_import_for]{pipeline_import_for()}},
\code{\link[=pipeline_init]{pipeline_init()}},
\code{\link[=pipeline_render_report]{pipeline_render_report()}},
\code{\link[=pipeline_run]{pipeline_run()}},
\code{\link[=pipeline_status]{pipeline_status()}},
\code{\link[=setup_environment]{setup_environment()}}
}
\concept{Pipeline functions}
