% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{read_non_empty_conf}
\alias{read_non_empty_conf}
\title{read non empty conf}
\usage{
read_non_empty_conf(f, cli_f, .call = parent.frame())
}
\arguments{
\item{f}{file location of config file}

\item{cli_f}{cli function to either warn or abort}

\item{.call}{Object: the parent environment from return_env, internal - do not use}
}
\description{
Helper function to read in a config if not empty
}
\keyword{internal}
