% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remove_comments}
\alias{remove_comments}
\title{remove comments}
\usage{
remove_comments(code)
}
\arguments{
\item{code}{String: code that needs to be stripped from comments}
}
\value{
String: input without comments
}
\description{
Helper function, using R's built-in parser to remove comments from code (both full- & inline)
Used instead of regex due to # usage in strings
e.g. my_string <- "using # inside string"
}
\keyword{internal}
