% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_cache.R
\name{save_cache}
\alias{save_cache}
\title{Save an object to the cache directory}
\usage{
save_cache(object, object_name, cache_dir)
}
\arguments{
\item{object}{The object that needs to be saved.}

\item{object_name}{The name of the object. This will be used as filename.}

\item{cache_dir}{The location of the cache. It is advised to use a timestamped directory.}
}
\description{
Saves an object to the cache directory. This function is dependent on DARA's directory structure.

save_cache is deprecated as of version 0.6.0.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
save_cache(cars, "cars", "cache/19000101_0000")
}
}
\seealso{
Other Save functions:
\code{\link[=save_data]{save_data()}},
\code{\link[=save_output]{save_output()}}
}
\concept{Save functions}
