% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_run.R
\name{save_image_offline}
\alias{save_image_offline}
\title{Save image offline}
\usage{
save_image_offline(object, bestand, device, ...)
}
\arguments{
\item{object}{Object to be saved as image}

\item{bestand}{File name to create on disk}

\item{device}{Device to use. Parameter to be used within ggsave}
}
\description{
Helper function that saves an image for offline publications.
}
\keyword{internal}
