% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_run.R
\name{save_object}
\alias{save_object}
\title{Save object}
\usage{
save_object(object_name, p_e, call = parent.frame())
}
\arguments{
\item{object_name}{Character: the name of the object}

\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\description{
Helper function that saves the object.
}
\keyword{internal}
