% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{set_hierarchy}
\alias{set_hierarchy}
\title{Set hierarchy}
\usage{
set_hierarchy(gr, call = parent.frame())
}
\arguments{
\item{gr}{Graph: Grab object that holds DAG}

\item{call}{the parent environment from return_env, internal - do not use}
}
\value{
hierarchy_level
}
\description{
Helper function to set hierarchy for \code{conf_create_dag}
}
\keyword{internal}
