% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_run.R
\name{source_object}
\alias{source_object}
\title{Source object}
\usage{
source_object(object_name, p_e, return_env = .GlobalEnv, call = parent.frame())
}
\arguments{
\item{object_name}{Character: the name of the object}

\item{p_e}{Object: package_environment, internal - do not use}

\item{return_env}{Object: the function execution environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\description{
helper function that sources the target objects.
}
\keyword{internal}
