% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{str_glue_expliciterror}
\alias{str_glue_expliciterror}
\title{string glue explicit error}
\usage{
str_glue_expliciterror(..., sep = "", .envir = parent.frame())
}
\arguments{
\item{...}{Additional arguments. Should be empty.}

\item{sep}{String: Seperator}

\item{.envir}{Object: the parent environment from return_env, internal - do not use}
}
\value{
Error
}
\description{
Helper function, used for wrapping explicit errors
used in: conf_register_object
}
\keyword{internal}
