% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_successful_run.R
\name{check_last_successful_run}
\alias{check_last_successful_run}
\title{Check last Successful run}
\usage{
check_last_successful_run(
  log_dir = "",
  include_cronjob = FALSE,
  cronjob_only = FALSE,
  search_window = 1,
  search_line = "",
  p_e = pipeline_env,
  call = parent.frame()
)
}
\arguments{
\item{log_dir}{string: Full path to log directory of desired project.}

\item{include_cronjob}{boolean: Include the logs from the cronjobs?}

\item{cronjob_only}{boolean: Only check cronjob logs? Default = FALSE, overwrites include_cronjob.}

\item{search_window}{integer: number of last lines in the logfile to check for. Default = 1}

\item{search_line}{string: Custom message to search for in log file}

\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\value{
timestamp of latest successful production run
}
\description{
Function that determines the last successful production run of the pipeline, based on log files.
}
\examples{
\dontrun{
Please note that the user has to provide the correct log directory path.
Find latest run, including cronjobs:
check_last_successful_run(log_dir = "Path_to_log_dir",
                          include_cronjob = TRUE)

Find latest run, excluding cronjobs:
check_last_successful_run(log_dir = "Path_to_log_dir",
                          include_cronjob = FALSE)

Look for custom completion messages:
check_last_successful_run(log_dir = "Path_to_log_dir"
                          include_cronjob = TRUE,
                          search_line = "Infectieradar saved clean joined, weekly and intake data.")

Look for messages that might not be on the last line of the log file:
check_last_successful_run(log_dir = "Path_to_log_dir",
                          include_cronjob = FALSE,
                          search_line = "Done!",
                          search_window = 2)
}

}
