% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{check_usage_dependencies}
\alias{check_usage_dependencies}
\title{Check usage dependencies}
\usage{
check_usage_dependencies(object_name, p_e = pipeline_env)
}
\arguments{
\item{object_name}{Character. The name of the object.}

\item{p_e}{pipeline environment}
}
\value{
invisible NULL
}
\description{
From the initialized pipeline environment, \code{check_usage_dependencies} checks whether
the dependencies specified in the configuration files for this object are being
used in the R object script. The location of the R script is retrieved from the
pipeline environment object \code{p_e}. The usage check is done by performing a string
search within the R script file with filename \code{object_name}.
}
\keyword{internal}
