% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_dependency_list.R
\name{copy_asset}
\alias{copy_asset}
\title{Copy single asset}
\usage{
copy_asset(
  data_asset_name,
  p_e = pipeline_env,
  call = parent.frame(),
  test_mode = FALSE
)
}
\arguments{
\item{data_asset_name}{Character. Name of object, as defined in object_definitions.yaml.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}

\item{test_mode}{For internal use. Boolean to inject \code{test_path()}
into file paths for testing purposes. Default = FALSE}
}
\description{
This functions copies an object/data to a given location under \code{copy_to_file_path} in object_definitions.yaml.
}
