% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DARApipeline-package.R
\docType{package}
\name{DARApipeline-package}
\alias{DARApipeline}
\alias{DARApipeline-package}
\title{DARApipeline: Tools To Run Your EPI Pipelines}
\description{
DARApipeline is an R package that provides you with the tools to run an EPI pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://dara.gitpages.rivm.nl/DARApipeline/index.html}
}

}
\author{
\strong{Maintainer}: DARAteam \email{dara-team-list@rivm.nl}

}
\keyword{internal}
