% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_from_dependency_list.R
\name{get_copy_from_file_path}
\alias{get_copy_from_file_path}
\title{Get the copy from file path from object_definitions}
\usage{
get_copy_from_file_path(
  data_asset_name,
  p_e,
  call = parent.frame(),
  test_mode = FALSE
)
}
\arguments{
\item{data_asset_name}{Character.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\value{
string with filepath to copy from
}
\description{
This functions uses the output_dir of a data_asset to determine where the object can be copied from.
}
\keyword{internal}
