% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_successful_run.R
\name{get_file_list}
\alias{get_file_list}
\title{helper get_file_list}
\usage{
get_file_list(log_dir, incl_cron, cron_only)
}
\arguments{
\item{log_dir}{string: path to the log directory}

\item{incl_cron}{boolean: Include cronjob logs}

\item{cron_only}{boolean: Only check cronjob logs? Default = FALSE, overwrites include_cron.}
}
\value{
list of log filepaths
}
\description{
helper get_file_list
}
\keyword{internal}
