% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_data.R
\name{get_sp_data}
\alias{get_sp_data}
\title{Retrieves data from a store procedure database within a start and end
modification date}
\usage{
get_sp_data(
  con_specs,
  start_date = "19000101",
  end_date = grab_run_timestamp(),
  answer_format = c("text", "code", "both")
)
}
\arguments{
\item{con_specs}{List. Connection arguments for the database. It should be a
named list with the names "con_args"/"sp_args". con_args should be a named
character with entries "server"/"database"/"uid"/"pwd" and sp_args should be
a named character with the entries "sp_name"/"infection_code".}

\item{start_date}{Date. Records retrieved from the the database will not
preceed this date. May also be a datetime or character string in the form
YYYYMMDD or YYYYMMDD_hhmm. Default = "19000101".}

\item{end_date}{Date. Records retrieved from the the database will not exceed
this date. May also be a datetime or character string starting with
YYYYMMDD. Default = run_timestamp.}

\item{answer_format}{Character. Provide questionare answer formats with
"text", "code" or "both".}
}
\value{
A tibble
}
\description{
get_sp_data() imports from the sp database with an
infection_code, start_ and end_date.
}
\examples{
\dontrun{
```
#### yaml password file example
connections:
sp_inf_disease: # as example, shape of yaml without using variables:
  con_args:
    server: "YOUR_SERVER"
    database: "OWH_DWN_HIS"
    uid: "USERNAME"
    pwd: "PASSWORD"
  sp_args:
    sp_name: "DWH_EPIRES"
    infection_code: "INF_CODE"
```
# import the above yaml
passwords <- read_yaml('passwords.yaml')

# 'jump back' to august 31st and retrieve data from august 1 to 31
data_legi <- get_sp_data(
  con_specs = config$connections$sp_inf_disease
  start_date  =  "20230801",
  end_date  =  "20230831"
)

}

}
