% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_object_cache.R
\name{grab_object_cache}
\alias{grab_object_cache}
\title{grab the cache path from the object_table in the pipeline_env}
\usage{
grab_object_cache(object_name, ..., p_e = pipeline_env, call = parent.frame())
}
\arguments{
\item{p_e}{Object: package_environment, internal - do not use}

\item{call}{Object: the parent environment from return_env, internal - do not use}
}
\description{
Function grabs the cache path from the object_table in the pipeline_env. Especially
useful when the cache file differs from the standard format cache/\{RUN_TIMESTAMP\}
}
\keyword{internal}
