% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_object_table.R
\name{grab_object_table}
\alias{grab_object_table}
\title{Grab object_params for all data_assets for the current EPI pipeline.}
\usage{
grab_object_table(..., p_e = pipeline_env, call = parent.frame())
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\value{
A tibble with object_params.
}
\description{
This function grabs the data_asset configuration (i.e. the object_params)
from the hidden pipeline state environment. These hold information such as
the paramaterised paths for retrieving / storing the data_asset.
}
