% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_run_timestamp.R
\name{grab_run_timestamp}
\alias{grab_run_timestamp}
\title{Grab the run_timestamp for the current EPI pipeline.}
\usage{
grab_run_timestamp(..., p_e = pipeline_env, call = parent.frame())
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\value{
A string in the form YYYMMDD_HHMM
}
\description{
Grabs and returns the used run_timestamp.
}
