% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_target_list.R
\name{grab_target_list}
\alias{grab_target_list}
\title{Grab a list of target objects given tag/object filters.}
\usage{
grab_target_list(
  tags = NULL,
  objects = NULL,
  ...,
  p_e = pipeline_env,
  call = parent.frame()
)
}
\arguments{
\item{tags}{Character. Which tags should be in the list of targets? Defaults
to all tags if objects = NULL.}

\item{objects}{Character. Which objects should be in the list of targets?
Defaults to all objects if tags = NULL.}

\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\value{
A character of object names.
}
\description{
Retrieves a list of object(s) given a tag or object filters. This
function is run by \link[DARApipeline]{pipeline_run} internally to determine which target
objects need to be created. It determines which other objects need to be generated in
case of dependencies between objects and returns these even if they fall
outside the tags/objects filter (because they are upstream of the pipeline).
Use this function to figure out which objects \link[DARApipeline]{pipeline_run} will create
without actually running \link[DARApipeline]{pipeline_run} yet.

Note that it only returns \strong{objects} not other types of data_assets
(data/other).
}
