% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_target_mount.R
\name{grab_target_mount}
\alias{grab_target_mount}
\title{Grab the mount for the current EPI pipeline.}
\usage{
grab_target_mount(..., p_e = pipeline_env, call = parent.frame())
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\value{
A string in the form of "rivm"
}
\description{
Grabs and returns the current mount for the EPI pipeline.
}
