% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idle.R
\name{idle}
\alias{idle}
\title{Idle}
\usage{
idle(
  dir,
  file_pattern,
  max_mins = 15,
  sleep_mins = 0.2,
  file_modification_buffer_mins = 1
)
}
\arguments{
\item{dir}{Character. Directory path.}

\item{file_pattern}{Character. Pattern of file that will be waited on.}

\item{max_mins}{Integer. Maximum number of minutes to be idle.}

\item{sleep_mins}{Double. Refresh time to check for file.}

\item{file_modification_buffer_mins}{Integer. Time since last update time of file.}
}
\description{
Waits for a file to be available in a given directory.
}
\examples{
\dontrun{
idle(".", "\.", max_mins = 1, sleep_mins = .2, file_modification_buffer_mins = 1)
}

}
