% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_config.R
\name{init_config}
\alias{init_config}
\title{Initialize config}
\usage{
init_config(dir_config, p_e, .call = parent.frame())
}
\arguments{
\item{dir_config}{Path: to directory which holds the configuration files.}

\item{p_e}{Object: package_environment, internal - do not use}

\item{.call}{Object: the parent environment from return_env, internal - do not use}
}
\description{
Initialize config
}
\section{Helper function kicked of in pipeline_init}{
This function was separated from the init.R file because of its size and many sub-helper functions
}

\keyword{internal}
