% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_init.R
\name{init_logging}
\alias{init_logging}
\title{Initialize logging}
\usage{
init_logging(path_logs, p_e)
}
\arguments{
\item{path_logs}{Path: Path to log file}

\item{p_e}{Object: package_environment, internal - do not use}
}
\description{
Helper function that intializes logging
}
\keyword{internal}
