% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data from specified location and date}
\usage{
load_data(loc, date, timerange, readFunc, fstcolumns = NULL, fstfilter = NULL)
}
\arguments{
\item{loc}{Character. Location of data (recursive).}

\item{date}{Date. Date in filename of the data in format "\%Y\%m\%d".}

\item{timerange}{Integer vector indicating time range}

\item{readFunc}{Function, with what function should data be read?}

\item{fstcolumns}{Character vector. Columns to load.}

\item{fstfilter}{Named list with 1 call.}
}
\value{
Most recent file with specified date in the name at specified location.
}
\description{
Load data from specified location and date.
Optionally a time range can be given and a file reading function.
When no file reading function is given, it is derived from the file extension.
With the fst extension, optional parameters can be given to filter the data.
}
\examples{
\dontrun{
load_data(loc = "./example", date = as.Date("2000-01-01"))
}

}
