% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_successful_run.R
\name{log_from_config}
\alias{log_from_config}
\title{Helper function to get log directory from config file}
\usage{
log_from_config(dir)
}
\arguments{
\item{dir}{string: current working directory.}
}
\value{
string: log directory path.
}
\description{
Helper function to get log directory from config file
}
\keyword{internal}
