% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_recent_file.R
\name{most_recent_file}
\alias{most_recent_file}
\title{Return path to most recent file in a directory with given pattern and
extension}
\usage{
most_recent_file(
  dir,
  pattern = ".",
  ext = "",
  show_found_files = FALSE,
  ignore_tmp = FALSE,
  call = parent.frame()
)
}
\arguments{
\item{dir}{Character. Path to directory.}

\item{pattern}{Character.}

\item{ext}{Character. Extension of the file.}

\item{show_found_files}{Logical.}

\item{ignore_tmp}{Logical. If TRUE, temporary owner files (~$) are ignored.}

\item{call}{Context of the parent interpreted function: environments
('frames' in S terminology) associated with functions further up the calling stack.}
}
\description{
Return path to most recent file in a directory with given pattern and
extension.
}
\examples{
\dontrun{
most_recent_data(dir = "./example")
}

}
