% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_successful_run.R
\name{order_file_list}
\alias{order_file_list}
\title{Helper order file list}
\usage{
order_file_list(file_list)
}
\arguments{
\item{file_list}{list: list of file paths to order by modification time}
}
\value{
df: containing file paths and their modification times, ordered by most recent
}
\description{
Helper order file list
}
\keyword{internal}
