% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_import_for.R
\name{pipeline_import_for}
\alias{pipeline_import_for}
\title{pipeline import for}
\usage{
pipeline_import_for(
  data_assets,
  ...,
  p_e = pipeline_env,
  call = parent.frame()
)
}
\arguments{
\item{data_assets}{Character. Name of the data_assets (objects/other) that you want to import objects for.
Multiple asset names should passed in an array of the form 'c()'.}

\item{...}{Additional arguments. Should be empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}

\item{call}{For internal use. Used for error tracing.}
}
\description{
Imports the dependencies (objects / data / other) given a
list of asset names. The imported objects will be loaded into the global
environment.
}
