% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_init.R
\name{pipeline_init}
\alias{pipeline_init}
\title{Initiate DARApipeline for an EPI pipeline}
\usage{
pipeline_init(run_timestamp = NULL, ..., p_e = pipeline_env)
}
\arguments{
\item{run_timestamp}{Character. run_timestamp in the format YYYMMDD_HHMM.
Defaults to the current day + time. Set this parameter to an old
run_timestamp to 'go back in time'.}

\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}
}
\description{
Initiates configuration of DARApipeline for use in an EPI pipeline. Does the following steps:
\enumerate{
\item Creates a run_timestamp
\item Initiates logging to log file
\item Interprets the configuration in config/
}

DARApipeline will keep track any parameters (such as the \code{run_timestamp} and the
configuration) in a hidden pipeline state environment.
}
\examples{
\dontrun{
pipeline_init()
}

}
\seealso{
\code{\link[=grab_object_table]{grab_object_table()}}.
}
