% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_status.R
\name{pipeline_status}
\alias{pipeline_status}
\title{Print a status overview of the EPI pipeline}
\usage{
pipeline_status(..., p_e = pipeline_env)
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}
}
\description{
This function prints the current status of the pipeline:
\enumerate{
\item Whether the pipeline has been initiated.
\item The initiated run_timestamp.
\item How many objects have been created so far.
\item The location of the log file.
}
}
