% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_vis.R
\name{pipeline_vis}
\alias{pipeline_vis}
\title{Visualizes object creation as a graph.}
\usage{
pipeline_vis(..., p_e = pipeline_env)
}
\arguments{
\item{...}{For internal use. Leave empty.}

\item{p_e}{For internal use. The hidden pipeline state environment.}
}
\description{
Visualizes the network defined by the object_relations yaml config file using
the \href{https://datastorm-open.github.io/visNetwork/}{visNetwork} package. This
function is meant for interactive use to view the structure of an EPI report
pipeline and can be very useful for inspecting the dependencies of your
objects. The pipeline_vis() function can maximally visualize 18 tags.
}
