% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_successful_run.R
\name{read_log}
\alias{read_log}
\title{Read log files}
\usage{
read_log(path, window, search_line)
}
\arguments{
\item{path}{string: path to the log files}

\item{window}{integer: specifies the size of the search window in the log files.}

\item{search_line}{string: Custom message to search for in log file}
}
\value{
success_file string of latest filename that successfully ran.
}
\description{
Read log files
}
\keyword{internal}
