% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Save data}
\usage{
save_data(
  object,
  file_directory,
  file_name,
  time = c(0, 2359),
  days = 1:7,
  ext = "rds",
  save_function = saveRDS,
  save_function_options = list(),
  run_timestamp = get("run_timestamp", envir = parent.frame()),
  timestamp_file_name = TRUE
)
}
\arguments{
\item{object}{R object to save.}

\item{file_directory}{Character. the location used for saving data.}

\item{file_name}{Character. (optional) name of the file to save, default = object.}

\item{time}{Numeric. Valid time range to save file(s) c(START HHMM, END HHMM),
example c(0000, 2359). Default = c(0000, 2359)}

\item{days}{Numeric. Valid days to save, 1 is monday and 7 is sunday, example for mo-fr: 1:5. Default = 1:7}

\item{ext}{Character. File extention. Default = "rds"}

\item{save_function}{Function. Optional function to use for saving. First 2 arguments need to be:
object, file location. Default = saveRDS.}

\item{save_function_options}{Character. Optional parameters for the save_function.}

\item{run_timestamp}{Character. YYYYMMDD_HHMM timestamp of the run.}

\item{timestamp_file_name}{Boolean. TRUE/FALSE do you want a timestamp in the file_name. Default = TRUE.}
}
\description{
Saves R objects as a file with a timestamp in a specified directory
}
\examples{
\dontrun{
save_data(
  object = cars,
  file_directory = "output/objects",
  file_name = "cars",
  run_timestamp = "19000101_0000"
)
}

}
