% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{setup_environment}
\alias{setup_environment}
\title{Setup the R environment for creation of objects}
\usage{
setup_environment(
  cur_env,
  prepare_script = "scripts/00_prepare/prepare.R",
  config_checks = TRUE,
  p_e = pipeline_env
)
}
\arguments{
\item{cur_env}{Environment. The environment from which setup_environment() is
called. object_name (only if cur_env == .GlobalEnv) and the list of
imported objects will be attached to this environment.}

\item{prepare_script}{Character. Location of the script that prepares R by
importing packages etc. Default ="scripts/00_prepare/prepare.R"}

\item{config_checks}{run the config checks for this object. Default is TRUE.}

\item{p_e}{pipeline environment.}
}
\value{
invisible(NULL)
}
\description{
setup_environment() is run at the top of an object creation
script and ensures the dependencies are loaded. If this function is run
interactively() by the user at the top of an object script, then,
additionally, setup_environment() deduces which object is worked on and
detemines the object_name.
}
\examples{
\dontrun{
setup_environment(environment())
### ### ### DO NOT EDIT ABOVE THIS LINE! ### ### ###
# (place the above at the top of a new object script)
}

}
